quran router removed  and surah router added
