+++ src/routers/quran/mod.rs
@@ -1,4 +1,3 @@
 pub mod mushaf;
-pub mod quran;
 pub mod surah;
 mod test;
